requirements.txt
docling
anthropic
pydantic
pandas
python-dotenv
